Add systemtests for basic coreutils
